refactor(products): 이벤트별 publisher 등록
