--- src/main/java/kitchenpos/common/Events.java
@@ -1,23 +1,0 @@
-package kitchenpos.common;
-
-import org.springframework.context.ApplicationEventPublisher;
-
-public class Events {
-
-    static ApplicationEventPublisher publisher;
-
-    public static void setPublisher(ApplicationEventPublisher publisher) {
-        if (Events.publisher == null) {
-            Events.publisher = publisher;
-        }
-    }
-
-    public static void raise(Object event) {
-        if (Events.publisher != null) {
-            publisher.publishEvent(event);
-        }
-    }
-
-    private Events() {
-    }
-}
