+++ src/main/java/kitchenpos/products/tobe/domain/events/ProductPriceChangedEvent.java
@@ -1,4 +1,4 @@
-package kitchenpos.products.tobe.domain;
+package kitchenpos.products.tobe.domain.events;
 
 import java.util.UUID;
